Automatically get remote for Pkg.add(...) packages
This will only work correctly for nicely tagged packages hosted on
GitHub. It works by getting hoping a tag v$VERSION exists, getting the
repository URL from the registry and the version number from the
manfiest. These tags are the type created by the widely used
JuliaRegistries/TagBot.

There is also a githubcheck() which uses the GitHub API to check the tag
exists and then check the package tree hash against the tree hash of the
corresponding commit.